+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- List controls in notebooks now include a per-item duplicate button.
+
 ## [0.13.3] - 2026-06-17
 
 ### Fixed